Rename utils file to Finskore
This is where I want to put all the game logic in a format
that will make it easy to use it in a non-vue app
and maintain game states on a server.
