+++ src/Finskore.js
